requirements.txt
fastapi
uvicorn
crewai
jinja2
aiofiles
requests
ollama               
elasticsearch
python-dotenv